Refactor: to the module structure I myself recommend. Heh.
